requirements.txt
wikipedia-api
bert-extractive-summarizer
torch